Compile StoreDataTypes
