+++ compiler/src/Language/Mimsa/Backend/Javascript/Printer.hs
@@ -7,6 +7,7 @@ module Language.Mimsa.Backend.Javascript.Printer
     printLiteral,
     printExpr,
     printStatement,
+    printDataType
   )
 where
